refactor(photos): group image zoom files
