+++ mobile/apps/photos/lib/ui/viewer/file/image_zoom/image_zoom_geometry.dart
